+++ CHANGELOG.md
@@ -77,3 +77,7 @@
 - New version 0.2.9
 - fix automatic restore database and store when update version 
 
+#### ⟳ Changelog Beta version 0.2.10 (2026-02-18)
+- Fix save key and value in project info tab.
+- Add new feacture, when delete key and value in project info tab, show a confirmation dialog.
+- New version 0.2.10 Beta